Better admin interface

+++ wsgiwars/views.py
@@ -39,32 +39,59 @@ def home(request):
 
 @view_config(route_name='delete_user')
 def delete(request):
-    #TODO if login
-    #TODO check is admin
+    if not 'login' in request.session:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
+    admin = request.session['login']
+    if not User.get(admin).is_admin:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
     user = User.get(request.matchdict['user'])
     user.delete()
     return HTTPFound(location=request.route_url('admin_list',page="0"))
 
 @view_config(route_name='admin')
 def admin(request):
+    if not 'login' in request.session:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
+    admin = request.session['login']
+    if not User.get(admin).is_admin:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
     return HTTPFound(location=request.route_url('admin_list',page="0"))
 
 @view_config(route_name='admin_list', renderer='templates/admin.pt')
 def admin_list(request):
-    #TODO if login
-    #TODO check is admin
+    if not 'login' in request.session:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
+    admin = request.session['login']
+    if not User.get(admin).is_admin:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
     skip = int(request.matchdict['page'])*10
     users = User.view('user/all', skip = skip , limit = 10, descending=True)
-    return {'users': users}
+    return {'users': users,
+            'page': request.matchdict['page']}
 
 @view_config(route_name='admin_user', renderer='templates/admin_user.pt')
 def admin_user(request):
-    #TODO if login
-    #TODO check is admin
+    if not 'login' in request.session:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
+    admin = request.session['login']
+    if not User.get(admin).is_admin:
+        request.session.flash(u"Sorry dude : Can't read this page")
+        return HTTPFound(location=request.route_url('home'))
     user = User.get(request.matchdict['user'])
     if request.method == 'POST':
         user.name = request.POST.get('name')
         user.description = request.POST.get('description')
+        if request.POST.get('admin') == 'on':
+            user.is_admin = True
+        else:
+            user.is_admin = False
         user.save()
     return {'user' : user}
 
@@ -105,6 +132,7 @@ def submitLogin(request):
     headers = remember(request, user._id)
     request.session['username'] = user.name
     request.session['login'] = user._id
+    request.session['is_admin'] = user.is_admin
     request.session.save()
 
     return HTTPFound(location=request.route_path('home'), headers=headers)
